WIP strncmp

+++ ft_strncmp.c
@@ -6,19 +6,19 @@
 /*   By: agaley <agaley@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/12 14:48:07 by agaley            #+#    #+#             */
-/*   Updated: 2022/11/09 20:12:36 by agaley           ###   ########lyon.fr   */
+/*   Updated: 2022/11/09 23:55:33 by agaley           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+// while ((s1[i] || s2[i]) && i < n)
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	// while ((s1[i] || s2[i]) && i < n)
 	while (n--)
 	{
 		if (*s1 != *s2 || *s1 == '\0' || *s2 == '\0')
-			return ((int)*s1 - (int)*s2);
+			return ((int)(*s1 - *s2));
 		s1++;
 		s2++;
 	}
